+++ src/lib.rs
@@ -12,7 +12,6 @@ use notify::RecursiveMode;
 use notify_debouncer_mini::new_debouncer;
 use std::sync::{mpsc, Arc, Mutex, Once};
 use std::{path::Path, time::Duration};
-use tokio::time;
 
 static INIT: Once = Once::new();
 
@@ -45,7 +44,7 @@ pub async fn run() -> Result<(), IotError> {
     client.run(None, methods, tx_client2app, rx_app2client);
 
     loop {
-        match rx_client2app.try_recv() {
+        match rx_client2app.recv_timeout(Duration::from_secs(RX_CLIENT2APP_TIMEOUT)) {
             Ok(Message::Authenticated) => {
                 INIT.call_once(|| {
                     #[cfg(feature = "systemd")]
@@ -92,7 +91,7 @@ pub async fn run() -> Result<(), IotError> {
             Ok(Message::C2D(msg)) => {
                 message::update(msg, Arc::clone(&tx_app2client));
             }
-            Err(mpsc::TryRecvError::Disconnected) => {
+            Err(mpsc::RecvTimeoutError::Disconnected) => {
                 return Err(IotError::from("iot channel unexpectedly closed by client"));
             }
             _ => {}
@@ -107,7 +106,5 @@ pub async fn run() -> Result<(), IotError> {
                 }
             })
         }
-
-        time::sleep(Duration::from_secs(RX_CLIENT2APP_TIMEOUT)).await;
     }
 }