+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.11] Q3 2022
+ - added direct method "reboot"
+ - fixed bug when async client does not terminate correctly
+ - improved logging for AuthenticationStatus changes
+
 ## [0.5.10] Q3 2022
  - log message with severity error on panics